рефакторинг такс-эдита + вынесли смарт класс в отдельный файл
